add socket-getting utils
